Create RemoveShortWords.cpp
I’ve been given a list of search terms (multiple lines, separated by spaces) that I need to comb through. All words of length 4 and shorter need to be removed.
Write a program that reads from standard input a series of search terms. This program should be in the file RemoveShortWords/main.cpp. Write to standard output only the words that are length 5 or greater, in the order of the original input separated by a single space character.